Move MemoryJmapSettingsRepository to the API module
