+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/settings/MemoryJmapSettingsRepository.scala
@@ -70,19 +70,3 @@ case class MemoryJmapSettingsRepository @Inject()() extends JmapSettingsReposito
     })
       .`then`()
 }
-
-case class MemoryJmapSettingsRepositoryModule() extends AbstractModule {
-  override def configure(): Unit = {
-    bind(classOf[MemoryJmapSettingsRepository]).in(Scopes.SINGLETON)
-    bind(classOf[JmapSettingsRepository]).to(classOf[MemoryJmapSettingsRepository])
-      .in(Scopes.SINGLETON)
-
-    Multibinder.newSetBinder(binder(), classOf[UsernameChangeTaskStep])
-      .addBinding()
-      .to(classOf[JmapSettingsUsernameChangeTaskStep])
-
-    Multibinder.newSetBinder(binder(), classOf[DeleteUserDataTaskStep])
-      .addBinding()
-      .to(classOf[JmapSettingsUserDeletionTaskStep])
-  }
-}
